+++ admb-server/src/authentication.ts
@@ -2,6 +2,17 @@ import e, { NextFunction, Request, Response } from "express";
 import { AuthenticationService } from "./services/authentication";
 import { Client } from "@metlife/appd-client";
 
+export const APPD_CLIENT = Symbol("APPD_CLIENT");
+export function client(req:Request):Client {
+   const client = (req as any)[APPD_CLIENT] as Client;
+   if (client) {
+      return client;
+   }
+   const error:any = new Error("Unauthorized")
+   error.status = 401;
+   throw error;
+}
+
 export const authenticationMiddleware = async (req:Request, rsp:Response, next:NextFunction) => {
    if (rsp.headersSent) {
       next();
@@ -10,11 +21,11 @@ export const authenticationMiddleware = async (req:Request, rsp:Response, next:N
    try {
       const session = await new AuthenticationService().authenticate(req);
       if (session) {
-         (req as any).client = new Client(session);
+         (req as any)[APPD_CLIENT] = new Client(session);
+         next();
       } else {
          rsp.sendStatus(401);
       }
-      next();
    } catch (e) {
       rsp
          .status(401)